Gen10/requirements.txt
shiny
python-dotenv
llama-index
llama-index-llms-gemini
llama-index-embeddings-huggingface
